Added checks on the WIF checksum.
When generating an address using a WIF private key, I've added a check to ensure that the checksum is correct. This is slightly more complicated, as it's possible someone generated a WIF using a full public key instead of a compressed private key, but that should be rare these days. I'll need to clean this up later to be safe though, and add some tests around it.
I added a quick unit test to ensure that a bad checksum fails to generate an address.