Updating P2SH imprementation.
The way I was building out the P2SH address was wrong. I updated it to be correct, and also updated the tests so that the library gives you the correct value.
I still need to fully implement P2SH so that it does more than encode a single key, and when I do that I'll come back and clean this up so it's moer readable. I'm spending most of my time on the f# version of this library now though.